PROJ_80 fix: brought out a DescriptionPost to share, and corrected comments
