+++ app/(app)/(withMenu)/profile/[userId]/page.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { ProfileUser, SelectedPost } from '@/features'
-import { EditPost } from '@/features/selectedPost/ui/editPost/EditPost'
+import { EditPostContainer } from '@/features/selectedPost/ui/editPost/EditPostContainer'
 import { useParams } from 'next/navigation'
 
 export default function ProfilePage() {
@@ -10,7 +10,7 @@ export default function ProfilePage() {
     <>
       <ProfileUser userId={userId as string} />
       <SelectedPost />
-      <EditPost />
+      <EditPostContainer />
     </>
   )
 }
