+++ features/createPost/api/api.ts
@@ -1,25 +1,8 @@
-import { BASE_URL_API } from '@/shared'
-import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
+import { baseApi } from '@/app'
 
 import { GetPosts, PostResponse, Publish } from './types'
 
-export const publishPostApi = createApi({
-  baseQuery: fetchBaseQuery({
-    baseUrl: BASE_URL_API,
-    credentials: 'include',
-    prepareHeaders: (headers, { getState }) => {
-      headers.set('User-Agent', navigator.userAgent)
-
-      const state: any = getState()
-      const token = state.auth?.accessToken
-
-      if (token) {
-        headers.set('Authorization', `Bearer ${token}`)
-      }
-
-      return headers
-    },
-  }),
+export const publishPostApi = baseApi.injectEndpoints({
   endpoints: builder => ({
     getNewestPosts: builder.query<PostResponse<Publish>, void>({
       query: () => 'posts/newest-posts',
@@ -34,7 +17,6 @@ export const publishPostApi = createApi({
       },
     }),
   }),
-  reducerPath: 'publishPostApi',
 })
 
 export const { useGetNewestPostsQuery, usePublishPostMutation } = publishPostApi
