+++ features/createPost/ui/steps/index.ts
@@ -1,4 +1,4 @@
 export * from './AddImage/AddImage'
 export * from './CropImage/CropImage'
 export * from './Filters/FiltersContainer'
-export * from './Publication/DescriptionPost'
+export * from './Publication/PublicationContainer'
