+++ features/selectedPost/ui/editPost/EditPost.module.scss
@@ -8,3 +8,7 @@
     bottom:  30px;
     right: 24px;
   }
+
+  .modal {
+    min-width: 500px;
+  }
