+++ features/selectedPost/ui/index.ts
@@ -1,1 +1,2 @@
 export * from './SelectedPost'
+export * from './editPost/EditPostContainer'
