--- shared/styles/index.ts
@@ -1,1 +1,0 @@
-export * from './DescriptionPost.module.scss'
