+++ shared/ui/DescriptionPost/DescriptionPost.tsx
@@ -1,12 +1,12 @@
-import React, { ChangeEvent, ReactNode, memo, useCallback } from 'react'
+import React, { ChangeEvent, ReactNode, useCallback } from 'react'
 
-import ava from '@/public/images/test_userAvatar.png'
 import { useAppDispatch } from '@/shared'
 import { TextArea } from '@rambo-react/ui-meteors'
 import { PayloadAction } from '@reduxjs/toolkit'
-import Image from 'next/image'
 
-import s from '@/shared/styles/DescriptionPost.module.scss'
+import s from './DescriptionPost.module.scss'
+
+import { NameAndAvatarUser } from './NameAndAvatarUser'
 
 type Props = {
   children: ReactNode
@@ -37,7 +37,7 @@ export const DescriptionPost = ({
   return (
     <div className={s.containerWrapper}>
       <div className={s.container}>
-        <ProfileInfo urlProfile={urlProfile} />
+        <NameAndAvatarUser urlProfile={urlProfile} />
         <TextArea
           className={s.description}
           label={'Add publication descriptions'}
@@ -49,15 +49,6 @@ export const DescriptionPost = ({
         {showSeparator ? <div className={s.separator} /> : null}
         {children}
       </div>
-    </div>
-  )
-}
-
-const ProfileInfo = ({ urlProfile }: { urlProfile: string }) => {
-  return (
-    <div className={s.avaWrapper}>
-      <Image alt={'userAvatar'} className={s.ava} src={ava} />
-      <span className={s.urlProfile}>{urlProfile}</span>
     </div>
   )
 }
