+++ shared/ui/DescriptionPost/NameAndAvatarUser.module.scss
@@ -1,0 +1,18 @@
+.ava {
+    height: 36px;
+    width: 36px;
+    border-radius: 50%;
+}
+
+.avaWrapper {
+    display: flex;
+    flex-direction: row;
+    margin-bottom: 24px;
+}
+
+.urlProfile {
+    padding-left:  12px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
