+++ shared/ui/index.ts
@@ -2,6 +2,7 @@ export * from './Carousel'
 export * from './ContentContainer'
 export * from './ControlledCheckbox'
 export * from './ControlledInput'
+export * from './DescriptionPost'
 export * from './ExpiredEmailLinkForm'
 export * from './FormContainer'
 export * from './NextNprogress'
